since an average is basicly a quotient, 'double' is a more logical type for it
git-svn-id: https://scm.mmbase.org/mmbase/trunk@36945 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ core/src/main/java/org/mmbase/cache/CacheMBean.java
@@ -70,7 +70,7 @@ public interface CacheMBean {
 
     int getByteSize();
 
-    int getAvarageValueLength();
+    double getAvarageValueLength();
